made minor refactoring
