+++ internal/repository/admin_repo.go
@@ -1,4 +1,4 @@
-package adminrepo
+package repository
 
 import (
 	"database/sql"
