+++ internal/repository/product_repo_test.go
@@ -1,4 +1,4 @@
-package productrepo
+package repository
 
 import (
 	"testing"
